src/TableAnalyzer/requirements.txt
pandas
statistics
xlsxwriter 
openpyxl
pyyaml